Add handling for updating end time based on playback speed
Adds handling to update the end time based on the user's current
playback speed.

This is a bit hacky since exoplayer is built on the assumption that we
have bi-directional communication to our UI layer. Instead the
CustomPlaybackTransportControlGlue is uni-directional polling the video
manager.

We unfortunately have to use hacky poll tricks to work-around the fact
we can't wire up on the "onPlaybackParametersChanged" call to notify the
glue class. However, in the short-term a 5s "delay" between setting and
updating is more favourable than setting up reverse listening
support.

Re #1509

+++ app/src/main/java/org/jellyfin/androidtv/ui/playback/VideoSpeedController.kt
@@ -3,16 +3,16 @@ package org.jellyfin.androidtv.ui.playback
 class VideoSpeedController(
 	private val parentController: PlaybackController
 ) {
-	enum class SpeedSteps(val speed: Double) {
+	enum class SpeedSteps(val speed: Float) {
 		// Use named parameter so detekt knows these aren't magic values
-		SPEED_0_25(speed = 0.25),
-		SPEED_0_50(speed = 0.5),
-		SPEED_0_75(speed = 0.75),
-		SPEED_1_00(speed = 1.0),
-		SPEED_1_25(speed = 1.25),
-		SPEED_1_50(speed = 1.50),
-		SPEED_1_75(speed = 1.75),
-		SPEED_2_00(speed = 2.0),
+		SPEED_0_25(speed = 0.25f),
+		SPEED_0_50(speed = 0.5f),
+		SPEED_0_75(speed = 0.75f),
+		SPEED_1_00(speed = 1.0f),
+		SPEED_1_25(speed = 1.25f),
+		SPEED_1_50(speed = 1.50f),
+		SPEED_1_75(speed = 1.75f),
+		SPEED_2_00(speed = 2.0f),
 	}
 
 	companion object {
